Renamed detail file.
